dependencies only used if optional feature active, refectored tlegen feature into separate module
